feat(config): support configurable window and pane numbering

+++ src/window.rs
@@ -3,7 +3,7 @@ use crate::pane::WindowId;
 #[derive(Debug, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
 pub struct WindowSummary {
     pub id: u64,
-    pub index: usize,
+    pub index: u64,
     pub name: String,
     pub active: bool,
     #[serde(default)]
@@ -13,7 +13,7 @@ pub struct WindowSummary {
 impl WindowSummary {
     pub fn new(
         id: WindowId,
-        index: usize,
+        index: u64,
         name: String,
         active: bool,
         last_selected: bool,
